pyproject.toml
[project]
name = "alexa-claude-bridge"
version = "0.2.0"
description = "Voice-control your running Claude Code REPL via Alexa + SQS"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35",
    "httpx>=0.28",
    "pywinpty>=2.0",
]

[project.scripts]
alexa-claude = "alexa_claude_bridge.wrapper:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alexa_claude_bridge"]

[tool.ruff]
target-version = "py313"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]
